modify models class
